+++ CHANGELOG.md
@@ -15,6 +15,9 @@ changes.
   path and noticeably increases sustained in-head throughput for script-heavy
   workloads. [2717](https://github.com/cardano-scaling/hydra/pull/2717)
 
+- Hydra node can now be configured through a yaml file; easier to spot
+  differences in configuration with peers. [#2296](https://github.com/cardano-scaling/hydra/issues/2296).
+
 ## [2.2.0] - 2026.06.12
 
 - Extend the end-to-end benchmark with real-world TPS metrics (end-to-end and